Decorator

+++ PlayGame.h
@@ -38,7 +38,7 @@ private:
     std::map<std::string, std::string> wordList;
     SendAnswer sendButton;
     sf::Vector2i mousePosition;
-    PlayerInput<int>* inputWord;
+    PlayerInput<unsigned short>* inputWord;
     sf::Text wordGuessed;
     int wordsGuessed=0;
     sf::Text wordDef;
